requirements_lock.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
chex==0.0.8
clang==5.0
clu==0.0.4
contextlib2==21.6.0
cycler==0.10.0
dill==0.3.4
dm-sonnet==2.0.0
dm-tree==0.1.6
flatbuffers==1.12
flax==0.3.4
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graph-nets==1.1.0
greenlet==1.1.1
grpcio==1.39.0
h5py==3.1.0
idna==3.2
--find-links https://storage.googleapis.com/jax-releases/jax_releases.html
jax==0.2.19
jaxlib==0.1.70
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
ml-collections @ git+git://github.com/google/ml_collections.git@7f749a281c69f9d0b339c05ecb94b80d95029f25
msgpack==1.0.2
neovim==0.3.1
networkx==2.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.0.9
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvim==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.15.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.6.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
toolz==0.11.1
tqdm==4.62.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
wheel==0.37.0
setuptools==57.4.0
python-version==0.0.2
jraph @ git+git://github.com/deepmind/jraph.git@88b962a1be3c93173247becf9814fa513ab87912
